+++ js/functions.js
@@ -20,3 +20,19 @@ checkPalindrome('ДовОд');
 checkPalindrome('Кекс');
 checkPalindrome('Лёша на полке клопа нашёл ');
 
+const hoursToMinutes = (time) => {
+  const timeAll = time.split(':');
+  return parseInt(timeAll[0],10) * 60 + parseInt(timeAll[1],10);
+};
+
+const workDay = (startWork, endWork, startMeet, timeMeet) => {
+  const startWorkMInutes = hoursToMinutes(startWork);
+  const stopWorkMinutes = hoursToMinutes(endWork);
+  const meetSum = hoursToMinutes(startMeet);
+
+  if (!(meetSum + timeMeet <= startWorkMInutes)) {
+    return (!(meetSum + timeMeet > stopWorkMinutes));
+  } return false;
+};
+
+workDay();
